Fixing old job cleanup https://www.wrike.com/open.htm?id=1251340096

+++ src/Async/TaskHandler/OldJobCleanupScheduledTaskHandler.php
@@ -1,4 +1,6 @@
-<?php declare(strict_types=1);
+<?php
+
+declare(strict_types=1);
 
 namespace Klaviyo\Integration\Async\TaskHandler;
 
@@ -11,20 +13,19 @@ use Shopware\Core\Framework\DataAbstractionLayer\EntityRepository;
 use Shopware\Core\Framework\DataAbstractionLayer\Search\Criteria;
 use Shopware\Core\Framework\DataAbstractionLayer\Search\Filter;
 use Shopware\Core\Framework\MessageQueue\ScheduledTask\ScheduledTaskHandler;
+use Shopware\Core\System\SystemConfig\SystemConfigService;
+use Symfony\Component\Messenger\Attribute\AsMessageHandler;
 
+#[AsMessageHandler(handles: OldJobCleanupScheduledTask::class)]
 class OldJobCleanupScheduledTaskHandler extends ScheduledTaskHandler
 {
-    private EntityRepository $jobRepository;
-    private LoggerInterface $logger;
-
     public function __construct(
         EntityRepository $scheduledTaskRepository,
-        EntityRepository $jobRepository,
-        LoggerInterface $logger
+        private readonly EntityRepository $jobRepository,
+        private readonly SystemConfigService $systemConfigService,
+        private readonly LoggerInterface $logger
     ) {
         parent::__construct($scheduledTaskRepository);
-        $this->jobRepository = $jobRepository;
-        $this->logger = $logger;
     }
 
     public static function getHandledMessages(): iterable
@@ -34,29 +35,34 @@ class OldJobCleanupScheduledTaskHandler extends ScheduledTaskHandler
 
     public function run(): void
     {
-        /*try {
-            // TODO: make it configurable from plugin settings in future
-            $numberOfDaysBeforeToday = new \DateTime(' - 5 day');
-            // Here we have context less task
-            $context = new Context(new SystemSource());
-            $criteria = new Criteria();
-            $criteria->addFilter(new Filter\AndFilter([
-                new Filter\RangeFilter(
-                    'createdAt',
-                    ['lt' => $numberOfDaysBeforeToday->format(Defaults::STORAGE_DATE_FORMAT)]
-                ),
-                new Filter\ContainsFilter('type', 'od-klaviyo'),
-                new Filter\EqualsFilter('parentId', null)
-            ]));
+        try {
+            $isJobCleanupEnabled = $this->systemConfigService->getInt('klavi_overd.config.oldJobCleanup');
+            $dayPeriod = $this->systemConfigService->getInt('klavi_overd.config.oldJobCleanupPeriod');
 
-            // Formatting IDs array and deleting config keys
-            $ids = \array_map(static function ($id) {
-                return ['id' => $id];
-            }, $this->jobRepository->searchIds($criteria, $context)->getIds());
+            if ($isJobCleanupEnabled && $dayPeriod) {
+                $numberOfDaysBeforeToday = new \DateTime(' - ' . $dayPeriod . ' day');
 
-            $this->jobRepository->delete($ids, $context);
+                // Here we have context less task
+                $context = new Context(new SystemSource());
+                $criteria = new Criteria();
+                $criteria->addFilter(new Filter\AndFilter([
+                    new Filter\RangeFilter(
+                        'createdAt',
+                        ['lt' => $numberOfDaysBeforeToday->format(Defaults::STORAGE_DATE_FORMAT)]
+                    ),
+                    new Filter\ContainsFilter('type', 'od-klaviyo'),
+                    new Filter\EqualsFilter('parentId', null),
+                ]));
+
+                // Formatting IDs array and deleting config keys
+                $ids = \array_map(static function ($id) {
+                    return ['id' => $id];
+                }, $this->jobRepository->searchIds($criteria, $context)->getIds());
+
+                $this->jobRepository->delete($ids, $context);
+            }
         } catch (\Throwable $e) {
             $this->logger->error($e->getMessage());
-        }*/
+        }
     }
 }
